Adding smoothing struct, fix-mag, weighted funct
Adding the smoothing function that returns the smoothing matrix
for fixed mag, weighted distribution label smoothing.

Adding the file structure in which other smoothing functions will
soon be implemented.

Adding test file, including two tests to be defined for the two
functions that are not yet defined, weighted-uniform and weighted-
adjacent smoothing. A weighting strategy has not yet been decided.